# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gorilla Guardians API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/checkout-session:
    post:
      operationId: createBookingCheckoutSession
      tags: [bookings]
      summary: Create a Stripe Checkout session to pay for an approved booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"

  # ORDERS
...
components:
  schemas:
...
    CheckoutSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Stripe-hosted Checkout page URL to redirect the browser to.

    # STORAGE
...
    Experience:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [guided_tour, homestay, workshop, dance_lesson, cooking_class]
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        price:
          type: number
        duration:
          type: string
        capacity:
          type: integer
        difficultyLevel:
          type: ["string", "null"]
        includedItems:
          type: array
          items:
            type: string
        meetingPoint:
          type: ["string", "null"]
        cancellationPolicy:
          type: ["string", "null"]
        active:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, description, type, images, price, duration, capacity, includedItems, active, reviewCount, createdAt]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        experienceId:
          type: integer
        date:
          type: string
          format: date
        participants:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        specialRequests:
          type: ["string", "null"]
        experience:
          $ref: "#/components/schemas/Experience"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, experienceId, date, participants, totalAmount, status, paymentStatus, createdAt]
    BookingInput:
      type: object
      required: [experienceId, date, participants]
      properties:
        experienceId:
          type: integer
        date:
          type: string
          format: date
        participants:
          type: integer
        specialRequests:
          type: string
        paymentMethod:
          type: string
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
        paymentStatus:
          type: string

    # ORDER
...